Update chocofi.keymap. Added td_numpadlayer to access numpad to both hands. This replaces the mirrored access to navigation keys and number row layers with the left and right hands respectively.
When using the left hand on the keyboard, the right hand is free to use the mouse. So, there is no need for the left hand to access the navigation keys. As for the right hand, there is no need for it to access the number row, because that can only be effectively used with both hands anyways. It would be better for each hands to have easy access to numpad, which is meant to be used by one hand.