Update chocofi.keymap. Added RGUI key on base layer of right handed side. This replaces right-handed numpad access.
Right-handed numpad access has been removed, because the numpad is more effectively used by the left; the left hand can be used to input all numbers by itself, the right hand is better of with the mouse. Moreover, when the right hand is not using the mouse, both hands would be available to use the number row anyways; this makes the right-handed numpad pointless.